re-organize code; only generate file if it doesn't exist
